add check_admin_referer&&wp_verify_nonce

+++ Simple-Post-Manager/Simple-Post-Manager.php
@@ -9,7 +9,6 @@
 	*/
 
 if ( ! defined( 'ABSPATH' ) ) exit; // Exit if access directly
-
 add_action('admin_menu', 'spm_SimplePostManager');
 
 // Here you can check if plugin is configured (e.g. check if some option is set). If not, add new hook. 
@@ -36,7 +35,8 @@ function spm_SimplePostManager_options() {
 					<th scope="row">Change all post to draft</th>
 					<td>
 						<form method="post">
-						<input type="submit" class="button-primary" name="postdraft" value="Apply"/>			
+							<?php  wp_nonce_field( 'postdraft_action', 'postdraft_nonce' ); ?>	
+							<input type="submit" class="button-primary" name="postdraft" value="Apply"/>			
 						</form>
 						All Post => (post_status=>'Draft')
 					</td>
@@ -45,7 +45,8 @@ function spm_SimplePostManager_options() {
 					<th scope="row">Delete All Post</th>
 					<td>
 						<form method="post">
-						<input type="submit" class="button-primary" name="postdelete" value="Apply"/>			
+							<input type="submit" class="button-primary" name="postdelete" value="Apply"/>
+							<?php wp_nonce_field( 'postdelete_action', 'postdelete_nonce' ); ?>		
 						</form>
 						This button deletes all post
 					</td>
@@ -55,41 +56,47 @@ function spm_SimplePostManager_options() {
 	</div>
 
 <?php
-
 	if( current_user_can( 'administrator' ) ){
-		
 		if(isset($_POST['postdraft'])){
+				if (check_admin_referer( 'postdraft_action', 'postdraft_nonce' ) && wp_verify_nonce($_POST['postdraft_nonce'], 'postdraft_action') ){
+			
 			$args = array('fields' => 'id=>post_status','numberposts' => -1);
 			$the_query = get_posts( $args );
-			if(!$the_query){
-				echo "no published post";
-			}
-			foreach($the_query as $post){
-			  $my_post = array();
-			  $my_post['ID'] = $post->ID;
-			  $my_post['post_status'] = 'draft';
-			  $ekle=wp_update_post( $my_post );
-			 if($ekle){echo "Successfully updated<br>";}	 
-			 else{echo "update failed";} 
-			}
-		}
+
+				if(!$the_query){echo "no published post";}
+					foreach($the_query as $post){
+					  $my_post = array();
+					  $my_post['ID'] = $post->ID;
+					  $my_post['post_status'] = 'draft';
+					  $ekle=wp_update_post( $my_post );
+
+						 if($ekle){echo "Successfully updated<br>";}	 
+						 else{echo "update failed";} 
+						 
+					}
+
+			    }//end check_admin...
+		}//end İsset($_POst..
+
 		if(isset($_POST['postdelete'])){
+			if (check_admin_referer( 'postdelete_action', 'postdelete_nonce' ) && wp_verify_nonce($_POST['postdelete_nonce'], 'postdelete_action') ){
+				$args = array('fields' => 'id=>post_status','numberposts' => -1);
+				$the_query = get_posts( $args );
+
+				if(!$the_query){echo "no published post";}
+
+				foreach($the_query as $post){
+				 $delete= wp_delete_post($post->ID,true);	
+
+					 if($delete){echo "Successfully deleted<br>";}	 
+					 else{echo "delete failed";}
+				}//end foreach
+
+			}//end if (check_admin_refe			
+		}//end if(isset($_POST['po
 
 
-			$args = array('fields' => 'id=>post_status','numberposts' => -1);
-			$the_query = get_posts( $args );
-			if(!$the_query){
-				echo "no published post";
-			}
-			foreach($the_query as $post){
-
-			 $delete= wp_delete_post($post->ID,true);
-			 if($delete){echo "Successfully updated<br>";}	 
-			 else{echo "delete failed";}
-			}				
-		}
-
-	}
+	}//end current_user_can
 ?>					
 	 <div class="wrap">
 		<div class="card pressthis">	
@@ -97,7 +104,7 @@ function spm_SimplePostManager_options() {
 		</div>
 	</div>									
 <?php
-}
+}//spm_SimplePostManager_options
 	
 ?>
